Some "TODO" have been removed, a missing check has been added in Interpol1D:linRegression (at least 2 valid points are required), type inconsistencies have been fixed in FilterTukey and ProcessinStack. However using size_t for the meteo parameter index is not necessarily a good idea and this will require a careful review.

+++ meteoio/meteofilters/ProcessingStack.cc
@@ -34,7 +34,7 @@ ProcessingStack::ProcessingStack(const Config& cfg, const std::string& parname) 
 		tmp << param_name << "::arg" << (ii+1);
 
 		getArgumentsForFilter(cfg, tmp.str(), vec_args); //Read arguments
-		filter_stack.push_back(BlockFactory::getBlock(block_name, vec_args));
+		filter_stack.push_back( BlockFactory::getBlock(block_name, vec_args) );
 	}
 }
 
@@ -102,7 +102,7 @@ void ProcessingStack::process(const std::vector< std::vector<MeteoData> >& ivec,
 					}
 					appliedFilter = true;
 
-					(*filter_stack[jj]).process(param, tmp, ovec[ii]);
+					(*filter_stack[jj]).process(static_cast<unsigned int>(param), tmp, ovec[ii]);
 
 					if (tmp.size() == ovec[ii].size()){
 						if ((jj+1) != filter_stack.size()){//after the last filter not necessary
